Just like on the ducks-server, got big changes here too:
Removed the next buttons and everything involving it. Was told it was tedious and annoying and honestly, they're kinda right. I thought it was cool at the time, but maybe it was just me being stupid. They're commented out instead of full-on deleted though, just in case I end up changing my mind, that happens sometimes.

For both the add and edit ducks feature, I changed the story text box so that it's bigger, because MY GOD, it was annoying to have to use the arrow keys for reading all the text I was putting in. I remember having to do that for some sites, still just as annoying as I remember, not going through that on my site.

For editing ducks, I made it so that you don't need to put in stuff for everything. Just put in what needs to be changed and the server will pass it on. Everything that's blank will remain unchanged.

Also, I FINALLY got started on putting the new features from the "allducks" collection to the other collections, starting with the "brand name ducks" collection. "Took me long enough," I know, but I am getting there.